Specify symfony/console version in composer update on GitHub Actions
This is necessary because the version of symfony/console will not be
updated correctly if it is not specified, although the cause is unknown.
If the version of symfony/console is not updated correctly, the error
"Some parallel worker jobs have not finished. while running parallel
worker" will occur when running PHPStan.